Added device multiplexer object
New device multiplexer object is introduced to manage the access  to a
single device from multiple processes.
For example when multiple virtual terminals access the keyboard, this
object will block most of the requests and forward only the active
stream. While the kernel will switch the active stream according to
which process is active.